refactor: move normalizer utilities into their own file
